pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fh_tailwindcss"
version = "0.0.3"
description = "Tailwind CSS package for FastHTML."
authors = [{ name = "lmbelo", email = "lucas.belo@live.com" }]
dependencies = ["python-fasthtml>=0.10.1"]
license = {file = "LICENSE"}
readme = {file="README.md", content-type="text/markdown"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["fh_tailwindcss*"]
exclude = ["samples"]    

[project.optional-dependencies]
flowbite = [
    "fh_flowbite>=0.0.1",
]